add build and install requirements to setup.py and pyproject.toml
The pyproject.toml file was added to express the build-time requirements
according to PEP 517 and 518. This became necessary, since we depend
on torch already at build time by importing torch in setup.py.
On top of that we filled the "install_requires" field in setup.py.

In principle pip is now itself able to resolve the dependencies during
installation, such that in a plain environment one could execute

	pip install .

However, build requirements are handled differently to install requirements.
Build requirments get installed into a temporary directory, and are only available
at build time. Thus listing torch as a build and install requirement leads to
torch being installed twice if one does a fresh install. To avoid this, installation
of the dependencies via the requirements.txt file will stay the recommended way.